Port to ABI3

+++ csrc/qattn/attn_cuda_sm90.h
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include <torch/extension.h>
+#include <torch/all.h>
 
 void qk_int8_sv_f8_accum_f32_block_sparse_attn_inst_buf_fuse_v_scale_sm90(
     torch::Tensor query,
@@ -26,10 +26,10 @@ void qk_int8_sv_f8_accum_f32_block_sparse_attn_inst_buf_fuse_v_scale_sm90(
     torch::Tensor query_scale,
     torch::Tensor key_scale,
     torch::Tensor value_scale,
-    int tensor_layout,
-    int is_causal,
-    int qk_quant_gran,
-    float sm_scale);
+    int64_t tensor_layout,
+    int64_t is_causal,
+    int64_t qk_quant_gran,
+    double sm_scale);
 
 torch::Tensor qk_int8_sv_f8_accum_f32_block_sparse_attn_inst_buf_fuse_v_scale_with_pv_threshold_sm90(
     torch::Tensor query,
@@ -42,8 +42,8 @@ torch::Tensor qk_int8_sv_f8_accum_f32_block_sparse_attn_inst_buf_fuse_v_scale_wi
     torch::Tensor query_scale,
     torch::Tensor key_scale,
     torch::Tensor value_scale,
-    int tensor_layout,
-    int is_causal,
-    int qk_quant_gran,
-    float sm_scale,
-    int return_pv_count);
+    int64_t tensor_layout,
+    int64_t is_causal,
+    int64_t qk_quant_gran,
+    double sm_scale,
+    int64_t return_pv_count);
